requirements.txt
Flask>=3.0.0
Pillow>=8.4.0
requests>=2.31.0
